+++ YourProjectWebApp/Models/Patron.cs
@@ -2,7 +2,7 @@
 {
     public class Patron
     {
-        public long PatronId { get; set; }
+        public int PatronId { get; set; }
         public string PatronName { get; set; }
         public bool IsGroup { get; set; }
     }
