+++ YourProjectWebApp/Models/Staff.cs
@@ -2,7 +2,7 @@
 {
     public class Staff
     {
-        public long StaffId { get; set; }
+        public int StaffId { get; set; }
         public string UserName { get; set; }
         public string UserPassword { get; set; }
     }
